Move Models into special folder
Now project separated on modules, which
contains all what they need for work correctly.
Little refactoring models.
